+++ Amazon.Pay.API.SDK.Tests/WebStore/Charge/CreateChargeRequestTests.cs
@@ -11,13 +11,13 @@ namespace Amazon.Pay.API.Tests.WebStore.Charge
         [Test]
         public void CanConstructWithAllPropertiesInitializedAsExpected()
         {
-            // arrange
+            // Arrange
             var chargePermissionId = "S02-7331650-8246451";
 
-            // act
+            // Act
             var request = new CreateChargeRequest(chargePermissionId, 12.99M, Currency.EUR);
 
-            // assert
+            // Assert
             Assert.IsNotNull(request);
             Assert.IsNotNull(request.ChargePermissionId);
             Assert.IsNotNull(request.ChargeAmount);
@@ -36,19 +36,19 @@ namespace Amazon.Pay.API.Tests.WebStore.Charge
         [Test]
         public void CanConvertToJsonMinimal()
         {
-            // arrange
+            // Arrange
             var chargePermissionId = "S02-7331650-8246451";
-            var request = new CreateChargeRequest(chargePermissionId, 12.99M, Currency.EUR);
+            var request = new CreateChargeRequest(chargePermissionId, 12.99M, Currency.EUR);  
+            // Act
+            string actualJson = request.ToJson();
+            
+            // Expected JSON string
+            string expectedJson = "{\"chargePermissionId\":\"S02-7331650-8246451\",\"chargeAmount\":{\"amount\":12.99,\"currencyCode\":\"EUR\"}}";
+            
+            // Assert
+            Assert.AreEqual(actualJson, expectedJson);
 
-            // act
-            string json = request.ToJson();
-            string json2 = request.ToJson();
-
-            // assert
-            Assert.AreEqual(json, json2);
-            Assert.AreEqual("{\"chargePermissionId\":\"S02-7331650-8246451\",\"chargeAmount\":{\"amount\":12.99,\"currencyCode\":\"EUR\"}}", json);
-
-            // verify object hasn't been corrupted
+            // Verify object hasn't been corrupted
             request.ProviderMetadata.ProviderReferenceId = "foo";
             request.SoftDescriptor = "foo";
             request.CaptureNow = true;
@@ -63,72 +63,109 @@ namespace Amazon.Pay.API.Tests.WebStore.Charge
         [Test]
         public void CanConvertToJsonFull()
         {
-            // arrange
+            // Arrange
             var chargePermissionId = "S02-7331650-8246451";
-            var request = new CreateChargeRequest(chargePermissionId, 12.99M, Currency.EUR);
-            request.ProviderMetadata.ProviderReferenceId = "foo";
-            request.SoftDescriptor = "foo";
-            request.CaptureNow = true;
-            request.PlatformId = "My Platform Id";
-            request.CanHandlePendingAuthorization = true;
+            var request = new CreateChargeRequest(chargePermissionId, 12.99M, Currency.EUR)
+                {
+                    ProviderMetadata = new ProviderMetadata { ProviderReferenceId = "foo1" },
+                    SoftDescriptor = "foo2",
+                    CaptureNow = true,
+                    PlatformId = "My Platform Id",
+                    CanHandlePendingAuthorization = true
+                };
 
-            // act
-            string json = request.ToJson();
-            string json2 = request.ToJson();
-
-            // assert
-            Assert.AreEqual(json, json2);
-            Assert.AreEqual("{\"chargePermissionId\":\"S02-7331650-8246451\",\"chargeAmount\":{\"amount\":12.99,\"currencyCode\":\"EUR\"},\"captureNow\":true,\"softDescriptor\":\"foo\",\"platformId\":\"My Platform Id\",\"canHandlePendingAuthorization\":true,\"providerMetadata\":{\"providerReferenceId\":\"foo\"}}", json);
+             // Act
+            string actualJson = request.ToJson();
+            
+            // Expected JSON string
+            string expectedJson = "{\"chargePermissionId\":\"S02-7331650-8246451\",\"chargeAmount\":{\"amount\":12.99,\"currencyCode\":\"EUR\"},\"captureNow\":true,\"softDescriptor\":\"foo2\",\"platformId\":\"My Platform Id\",\"canHandlePendingAuthorization\":true,\"providerMetadata\":{\"providerReferenceId\":\"foo1\"}}";
+            
+            // Assert
+            Assert.AreEqual(actualJson, expectedJson);
         }
 
         [Test]
         public void CanConvertToJsonRecurring()
         {
-            // arrange
+            // Arrange
             var chargePermissionId = "S02-7331650-8246451";
-            var request = new CreateChargeRequest(chargePermissionId, 12.99M, Currency.EUR);
-            request.ProviderMetadata.ProviderReferenceId = "foo1";
-            request.SoftDescriptor = "foo2";
-            request.CaptureNow = true;
-            request.PlatformId = "My Platform Id";
-            request.CanHandlePendingAuthorization = true;
-            request.MerchantMetadata.MerchantReferenceId = "123abc!";
-            request.MerchantMetadata.MerchantStoreName = "My Store Name";
-            request.MerchantMetadata.NoteToBuyer = "My Note to Buyer";
-            request.MerchantMetadata.CustomInformation = "My Custom Info";
+            var request = new CreateChargeRequest(chargePermissionId, 12.99M, Currency.EUR)
+                {
+                    ProviderMetadata = new ProviderMetadata { ProviderReferenceId = "foo1" },
+                    SoftDescriptor = "foo2",
+                    CaptureNow = true,
+                    PlatformId = "My Platform Id",
+                    CanHandlePendingAuthorization = true,
+                    MerchantMetadata = new MerchantMetadata 
+                        { 
+                            MerchantReferenceId = "123abc!",
+                            MerchantStoreName = "My Store Name",
+                            NoteToBuyer = "My Note to Buyer",
+                            CustomInformation = "My Custom Info",
+                        }
+                };
 
-            // act
-            string json = request.ToJson();
-            string json2 = request.ToJson();
-
-            // assert
-            Assert.AreEqual(json, json2);
-            Assert.AreEqual("{\"chargePermissionId\":\"S02-7331650-8246451\",\"chargeAmount\":{\"amount\":12.99,\"currencyCode\":\"EUR\"},\"captureNow\":true,\"softDescriptor\":\"foo2\",\"platformId\":\"My Platform Id\",\"canHandlePendingAuthorization\":true,\"providerMetadata\":{\"providerReferenceId\":\"foo1\"},\"merchantMetadata\":{\"merchantReferenceId\":\"123abc!\",\"merchantStoreName\":\"My Store Name\",\"noteToBuyer\":\"My Note to Buyer\",\"customInformation\":\"My Custom Info\"}}", json);
+            // Act
+            string actualJson = request.ToJson();
+            
+            // Expected JSON string
+            string expectedJson = "{\"chargePermissionId\":\"S02-7331650-8246451\",\"chargeAmount\":{\"amount\":12.99,\"currencyCode\":\"EUR\"},\"captureNow\":true,\"softDescriptor\":\"foo2\",\"platformId\":\"My Platform Id\",\"canHandlePendingAuthorization\":true,\"providerMetadata\":{\"providerReferenceId\":\"foo1\"},\"merchantMetadata\":{\"merchantReferenceId\":\"123abc!\",\"merchantStoreName\":\"My Store Name\",\"noteToBuyer\":\"My Note to Buyer\",\"customInformation\":\"My Custom Info\"}}";
+            
+            // Assert
+            Assert.AreEqual(actualJson, expectedJson);
         }
 
         [Test]
         public void CanConvertToJsonPaymentMethodOnFile()
         {
-            // arrange
+            // Arrange
             var chargePermissionId = "S02-7331650-8246451";
-            var request = new CreateChargeRequest(chargePermissionId, 12.99M, Currency.EUR);
-            request.ProviderMetadata.ProviderReferenceId = "foo";
-            request.SoftDescriptor = "foo";
-            request.CaptureNow = true;
-            request.PlatformId = "My Platform Id";
-            request.CanHandlePendingAuthorization = true;
-            request.ChargeInitiator = ChargeInitiator.CITR;
-            request.Channel = Channel.Web;
+            var request = new CreateChargeRequest(chargePermissionId, 12.99M, Currency.EUR)
+                {
+                    ProviderMetadata = new ProviderMetadata { ProviderReferenceId = "foo" },
+                    SoftDescriptor = "foo",
+                    CaptureNow = true,
+                    PlatformId = "My Platform Id",
+                    CanHandlePendingAuthorization = true,
+                    ChargeInitiator = ChargeInitiator.CITR,
+                    Channel = Channel.Web,
+                };
 
-            // act
-            string json = request.ToJson();
-            string json2 = request.ToJson();
+            // Act
+            string actualJson = request.ToJson();
+            
+            // Expected JSON string
+            string expectedJson = "{\"chargePermissionId\":\"S02-7331650-8246451\",\"chargeAmount\":{\"amount\":12.99,\"currencyCode\":\"EUR\"},\"captureNow\":true,\"softDescriptor\":\"foo\",\"platformId\":\"My Platform Id\",\"canHandlePendingAuthorization\":true,\"providerMetadata\":{\"providerReferenceId\":\"foo\"},\"chargeInitiator\":\"CITR\",\"channel\":\"Web\"}";
 
-            // assert
-            Assert.AreEqual(json, json2);
-            Assert.AreEqual("{\"chargePermissionId\":\"S02-7331650-8246451\",\"chargeAmount\":{\"amount\":12.99,\"currencyCode\":\"EUR\"},\"captureNow\":true,\"softDescriptor\":\"foo\",\"platformId\":\"My Platform Id\",\"canHandlePendingAuthorization\":true,\"providerMetadata\":{\"providerReferenceId\":\"foo\"},\"chargeInitiator\":\"CITR\",\"channel\":\"Web\"}", json);
+            // Assert
+            Assert.AreEqual(actualJson, expectedJson);
         }
 
+        [Test]
+        public void CanConvertToJsonSavedWallet()
+        {
+            // Arrange
+            var chargePermissionId = "S01-7436918-9739892";
+            var checkoutResultReturnUrl = "https://example.com/return.html";
+            var request = new CreateChargeRequest(chargePermissionId, 99.99M, Currency.EUR, checkoutResultReturnUrl)
+                {
+                    ProviderMetadata = new ProviderMetadata { ProviderReferenceId = "foo" },
+                    SoftDescriptor = "foo",
+                    CaptureNow = true,
+                    PlatformId = "My Platform Id",
+                    CanHandlePendingAuthorization = true,
+                    ChargeInitiator = ChargeInitiator.CITR,
+                    Channel = Channel.Web
+                };
 
+            // Act
+            string actualJson = request.ToJson();
+            
+            // Expected JSON string
+            string expectedJson = "{\"chargePermissionId\":\"S01-7436918-9739892\",\"chargeAmount\":{\"amount\":99.99,\"currencyCode\":\"EUR\"},\"captureNow\":true,\"softDescriptor\":\"foo\",\"platformId\":\"My Platform Id\",\"canHandlePendingAuthorization\":true,\"providerMetadata\":{\"providerReferenceId\":\"foo\"},\"chargeInitiator\":\"CITR\",\"channel\":\"Web\",\"webCheckoutDetails\":{\"checkoutResultReturnUrl\":\"https://example.com/return.html\"}}";
+
+            // Assert
+            Assert.AreEqual(actualJson, expectedJson);
+        }
     }
 }